fix: fixes (1) to run tests on hetzner-volterra
- make namespace where mayastor is deployed configurable
- add configurations for
    - setting HostNetwork:true for test pods
    - mayastor namespace
    - filtered mayastor pod health check  (only check mayastor pods in namespace)
    - inability to restore pools, we do not have a complete solution for this as yet
    - temporary config to disable e2e-agent
- hacky setting of the namspace value
- disable pool restoration dependent on configuration
- make cleanup of PVCs and PVs more accurate,
    i.e. only delete those provisioned using storage classes provisioning
    using mayastor.
- make checking of PVCs and PVs more accurate,
    i.e. only check those provisioned using storage classes provisioning
    using mayastor.
- replace some uses of fmt.Println
- Add HostNetwork:true in the spec if required by the configuration
- Workaround the fact the volterra messes pod completion by adding the wingman pod,
    look at container statuses instead IMHO this is more fragile.

+++ src/common/k8stest/util.go
@@ -410,17 +410,23 @@ func CreateConfiguredPools() {
 // As part of the tests we may modify the pools, in such test cases
 // the test should delete all pools and recreate the configured set of pools.
 func RestoreConfiguredPools() error {
-	logf.Log.Info("RestoreConfiguredPools: delete all pools and re-create configured pools.")
-	deletedAllPools := DeleteAllPools()
-	Expect(deletedAllPools).To(BeTrue())
-	CreateConfiguredPools()
-	const sleepTime = 5
-	for ix := 1; ix < 120/sleepTime; ix++ {
-		time.Sleep(sleepTime * time.Second)
-		err := CheckAllPoolsAreOnline()
-		if err == nil {
-			break
+	cannotRestorePools := e2e_config.GetConfig().CannotRestorePools
+	logf.Log.Info("RestoreConfiguredPools: delete all pools and re-create configured pools.", "cannotRestorePools", cannotRestorePools)
+	// FIXME: On some test deployments we are unable to recreate the pools yet.
+	if !cannotRestorePools {
+		deletedAllPools := DeleteAllPools()
+		Expect(deletedAllPools).To(BeTrue())
+		CreateConfiguredPools()
+		const sleepTime = 5
+		for ix := 1; ix < 120/sleepTime; ix++ {
+			time.Sleep(sleepTime * time.Second)
+			err := CheckAllPoolsAreOnline()
+			if err == nil {
+				break
+			}
 		}
+	} else {
+		logf.Log.Info("RestoreConfiguredPools: pools not restored, configuration does not support this feature.")
 	}
 	return CheckAllPoolsAreOnline()
 }
